Add automatic orientation.
This branch adds an automaticOrientation prop that controls the
positioning of the x and y axes.

As an experiment, it shows that automatic orientation is possible,
however, it operates against height and width which do not share a
common size. Since the height unit is so much smaller than the width
unit, the resulting axes positions look really odd.

If this was updated to factor in actual pixel sizes, it would work
appropriately.